webapps/djangofile/requirements.txt
Django==5.1.1
azure-identity==1.17.1
azure-storage-blob==12.22.0
aiohttp==3.10.5
